requirements.txt
cffi==1.4.1
cryptography==1.1.2
Django==1.8.7
django-mptt==0.7.4
enum34==1.1.1
idna==2.0
ipaddress==1.0.15
MySQL-python==1.2.5
ndg-httpsclient==0.4.0
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==0.15.1
six==1.10.0
wheel==0.24.0
